backend/requirements.txt
# Turgon — Autonomous Policy-to-Enforcement Engine
# Core Agent Framework
crewai>=0.80.0
crewai-tools>=0.14.0

# LLM / Groq client
langchain-groq>=0.1.0

# PDF Ingestion
docling>=2.3.0

# Database
duckdb>=1.1.0

# Frontend (Streamlit — kept for reference)
streamlit>=1.41.0
plotly>=5.20.0

# Flask REST API Backend
flask>=3.0.0
flask-cors>=4.0.0

# Vector DB
chromadb>=0.5.0

# Data & Validation
pandas>=2.2.0
pydantic>=2.6.3
sqlparse>=0.5.0

# Utilities
python-dotenv>=1.0.1
rich>=13.7.0
psutil>=5.9.0
reportlab>=4.0.0  # PDF generation for reports
